check for markdown dependency before further executing script

+++ generate_website.sh
@@ -1,11 +1,18 @@
 #!/bin/sh
 
+# check for dependencies
+markdown version > /dev/null
+if [[ $? != 0 ]]; then
+  echo "unmet dependency: markdown"
+  exit
+fi
+
 # ensure working directory is clean
 rm -r html 2>/dev/null
 mkdir html 2>/dev/null
 
 # move into markdown directory to make some path stuff easier
-cd markdown
+cd ./markdown
 
 # make directory structure
 for DIRECTORY in $(find . -mindepth 1 -type d)
